+++ Tests/DatadogTests/Datadog/LoggerTests.swift
@@ -27,7 +27,6 @@ class LoggerTests: XCTestCase {
     func testSendingLogWithDefaultLogger() throws {
         core.context = .mockWith(
             configuration: .mockWith(
-                applicationVersion: "1.0.0",
                 applicationBundleIdentifier: "com.datadoghq.ios-sdk",
                 serviceName: "default-service-name",
                 environment: "tests",
@@ -35,7 +34,8 @@ class LoggerTests: XCTestCase {
             ),
             dependencies: .mockWith(
                 dateProvider: RelativeDateProvider(using: .mockDecember15th2019At10AMUTC())
-            )
+            ),
+            appVersionProvider: .mockWith(version: "1.0.0")
         )
 
         let feature: LoggingFeature = .mockByRecordingLogMatchers()
@@ -452,9 +452,9 @@ class LoggerTests: XCTestCase {
     func testSendingTags() throws {
         core.context = .mockWith(
             configuration: .mockWith(
-                applicationVersion: "1.2.3",
                 environment: "tests"
-            )
+            ),
+            appVersionProvider: .mockWith(version: "1.2.3")
         )
 
         let feature: LoggingFeature = .mockByRecordingLogMatchers()