+++ Tests/DatadogTests/Datadog/TracerTests.swift
@@ -27,7 +27,6 @@ class TracerTests: XCTestCase {
     func testSendingSpanWithDefaultTracer() throws {
         core.context = .mockWith(
             configuration: .mockWith(
-                applicationVersion: "1.0.0",
                 applicationBundleIdentifier: "com.datadoghq.ios-sdk",
                 serviceName: "default-service-name",
                 environment: "custom",
@@ -36,7 +35,8 @@ class TracerTests: XCTestCase {
             ),
             dependencies: .mockWith(
                 dateProvider: RelativeDateProvider(using: .mockDecember15th2019At10AMUTC())
-            )
+            ),
+            appVersionProvider: .mockWith(version: "1.0.0")
         )
 
         let feature: TracingFeature = .mockByRecordingSpanMatchers(